Fix nested data causing 422 Unprocessable Entity
When you try to `$api->tags()->post(['data' => ...]);`, the array
passed gets nested, resulting in `['json' => [0 => ['data' => ...]]]`
which the API can't parse, as it expects data immediately after json.

This is caused by the `__call` magic method in Fluent.php and could
even be fixed there instead. Interestingly, that also passes the
$arguments variable to `setMethod()` which doesn't take or use it.

Without knowing for certain whether other methods take attributes
differently, this is probably a bit of a safer fix. Watch this space.